+++ packages/telemetry/index.d.ts
@@ -1,0 +1,4 @@
+/**
+ * typedefs for telemetry
+ */
+export * from './index.cjs';