+++ sites/test-env/CHANGELOG.md
@@ -1,5 +1,12 @@
 # evidence-test-environment
 
+## 3.0.6
+
+### Patch Changes
+
+- @evidence-dev/evidence@24.0.6
+- @evidence-dev/core-components@3.0.3
+
 ## 3.0.5
 
 ### Patch Changes